Renaming React
